requirements.txt
python-docx
fastapi==0.95.0
python-dotenv
uvicorn==0.21.1
python-multipart
